fixedup
Former-commit-id: 3af072b4c807804c2c3d30c9686ce6a6a3abf0d3

+++ themes/material/modules/goods_received_note/create.php
@@ -1025,7 +1025,7 @@
                 $('#received_unit_value').attr('readonly', true);
                 $('#purchase_order_number').attr('readonly', true);
 
-                $('#quantity_order').data('rule-max', parseInt(ui.item.left_received_quantity)).data('msg-max', 'max available ' + ui.item.left_received_quantity);
+                $('#quantity_order').data('rule-max', parseFloat(ui.item.left_received_quantity)).data('msg-max', 'max available ' + ui.item.left_received_quantity);
 
                 // if (ui.item.serial_number != null){
                 //   $( inputIssuedQuantity ).val(1).attr('readonly', true);
